dimos/models/manipulation/contact_graspnet_pytorch/requirements.txt

opencv-python==4.10.0.84
python-dotenv
openai
anthropic>=0.19.0
cerebras-cloud-sdk
numpy>=1.24.3,<2.0.0
colorlog==6.9.0
yapf==0.40.2
typeguard
empy==3.3.4
catkin_pkg
lark
plum-dispatch==2.5.7

# pycolmap
ffmpeg-python
pytest
python-dotenv
openai
tiktoken>=0.8.0
Flask>=2.2
python-multipart==0.0.20
reactivex
git+https://github.com/dimensionalOS/rxpy-backpressure.git
pytest-asyncio==0.26.0
asyncio==3.4.3
-e git+https://github.com/legion1581/go2_webrtc_connect.git@fe64abb5987594e8c048427a98445799f6f6a9cc#egg=go2_webrtc_connect
#-e git+https://github.com/legion1581/aioice.git@ff5755a1e37127411b5fc797c105804db8437445#egg=aioice

# Web Extensions
fastapi>=0.115.6
sse-starlette>=2.2.1
uvicorn>=0.34.0

# Agent Memory 
langchain-chroma>=0.1.4
langchain-openai>=0.2.14

# Class Extraction
pydantic

# Developer Specific
ipykernel

# Unitree webrtc streaming
aiortc==1.9.0
pycryptodome
sounddevice
pyaudio
requests
wasmtime

# Audio
openai-whisper
soundfile

#Hugging Face
transformers[torch]==4.49.0

#Vector Embedding
sentence_transformers

# CTransforms GGUF - GPU required
ctransformers[cuda]==0.2.27

# Perception Dependencies
ultralytics>=8.3.70
filterpy>=1.4.5
scipy>=1.10.1
opencv-contrib-python==4.10.0.84
scikit-learn>=1.2.2
Pillow
mmengine>=0.10.3
mmcv>=2.1.0
timm>=1.0.15
lap>=0.5.12 
xformers==0.0.20

# Detic
mss
dataclasses
ftfy
regex
fasttext
lvis
nltk
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/detectron2.git@v0.6

# Mapping
open3d>=0.17.0

# Contact Graspnet Dependencies--------------------
h5py>=3.7.0
pyrender>=0.1.45
trimesh>=3.22.0
python-fcl>=0.7.0.4
pyquaternion>=0.9.9
matplotlib>=3.7.1
rtree
pandas>=1.5.2
requests>=2.29.0
tqdm>=4.65.0
pyyaml>=6.0

# Visualization (Optional)
# mayavi>=4.8.1
kaleido>=0.2.1
# python-kaleido>=0.2.1
plotly>=5.9.0
# Contact Graspnet Dependencies--------------------

